Create C and NC as online and offline persistence service
