+++ Annotato/Annotato/Persistence/PersistenceManager/AnnotatoPersistence.swift
@@ -2,18 +2,20 @@ import Foundation
 import AnnotatoSharedLibrary
 
 struct AnnotatoPersistence {
-    static var currentPersistenceManager: PersistenceManager {
-        if WebSocketManager.shared.isConnected {
-            onlinePersistenceManager
-        } else {
-            offlinePersistenceManager
-        }
+    static var currentPersistenceService: PersistenceService {
+        WebSocketManager.shared.isConnected ? onlinePersistenceManager : offlinePersistenceManager
     }
+
     static let remotePersistence: PersistenceManager = RemotePersistenceManager()
     static let localPersistence: PersistenceManager = LocalPersistenceManager.shared
 
-    static let onlinePersistenceManager: PersistenceService = OnlinePersistenceManager(remotePersistence: remotePersistence, localPersistence: localPersistence)
-    static let offlinePersistenceManager: PersistenceService = OfflinePersistenceManager(localPersistence: localPersistence)
+    static let onlinePersistenceManager: PersistenceService = OnlinePersistenceService(
+        remotePersistence: remotePersistence,
+        localPersistence: localPersistence
+    )
+    static let offlinePersistenceManager: PersistenceService = OfflinePersistenceService(
+        localPersistence: localPersistence
+    )
 
     private init() { }
 }
